report: fix bug where some ipmeta-based series could have bad labels
Metric IDs that do not appear in the appropriate label map would
produce "empty" metrictype and metricval strings for constructing
the time series label.

An ID can be missing because:
 * the tagger was not configured to load the full tagging data for
   that particular metric, but was able to tag packets for that
   metric anyway (most likely to occur with netacq regions).
 * the corsarotrace instance is using a faux control socket but
   it wasn't configured with the same tagging information as the
   original tagger.
 * some other misconfiguration or bug that causes metric IDs to
   not match what we have labels for.

Now, any series where we are unable to map a metric ID to a
label from IPmeta will be quietly ignored rather than being
output with a near-useless and confusing label.