Separate public flowtuple structure from internal one
Previously, we were installing a header that nobody could actually
use as it pulled in other internal headers. However, it does
make sense for the flowtuple "data" to be defined as a structure
that is available in the public API (especially along with our
avro encode and decode functions).

Key changes:
 * added libcorsaro_flowtuple.h file which will be installed in
   $(prefix)/include -- this contains a flowtuple_data structure
   definiton and prototypes for the avro encode/decode functions.
 * updated flowtuple structure definition in
   plugins/corsaro_flowtuple.h to have a flowtuple_data member,
   as well as all the internal members (priority queue stuff,
   sorting keys etc).
 * plugins/*.h are now no longer installed by corsaro. Any code
   that was using these headers (and therefore probably didn't
   work) should now use libcorsaro_flowtuple.h instead.
 * updated corsaroftmerge to use the new data structure.