Disable tracking packet timestamps for dos attack vectors
The timestamps have never been used to generate any output, nobody
seems to notice that they are missing and the code to maintain
and merge the timestamp lists has a huge performance cost.

We can consider re-adding this feature (and writing it in a more
efficient way), IF someone actually has a genuine use case for it.