rsdos: count attacker /16s instead of individual IPs
Maintaining a map of unique IPs seen from the telescope can
require a lot of memory and CPU time if the DoS is long-running
and spoofs a large portion of the telescope address space.

To reduce the impact, we've decided that simply counting unique
/16s will give a reasonable idea of the range of addresses that
are being used for any given attack, while requiring far fewer
resources to track and maintain.

Individual attacks where a researcher is particularly interested
in the distribution of specific telescope addresses can still be
investigated directly using the raw pcaps.