Add container + scripts for mass converting rsdos2 data to simple ascii
The data is a pain for people to use as they need corsaro2 to be
able to parse the old files -- using this container, we can
easily convert the old files to a simple ASCII CSV format
and put those in a Swift container that users can access instead.

We only really use the "initial packet" that was recorded for
each attack, but (as always) this could be found in the pcaps
(or the rsdos2 cors files if they are kept) if someone really
wanted this.

Documentation pending, but this container works very much like
the other offline containers in this repo so shouldn't be too
hard to figure out...